+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 
 Unreleased
 ==========
-
+* feat: Enable add button to crate a snippet when adding a SnippetPlugin
 
 4.0.1.dev1 (2022-05-10)
 =======================